backend/requirements.txt
flask
flask-cors
langchain
langchain-groq
python-dotenv